backend/requirements.txt
Flask
Flask-Cors
scikit-learn
nltk
python-dotenv
google-generativeai
openai
